zebra: rib->nexthop_num is double incremented
While reading linux kernel netlink messages, rib->nexthop_num is incremented
in the read loop as well as the underlying nexthop_add() function call.  This
number should only be incremented in one spot.  Removing the increment from
the read loop and allowing the underlying code to track this data as it is
already.

Signed-off-by: Donald Sharp <sharpd@cumulusnetworks.com>